% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate via Github}
\usage{
authenticate(user, password)
}
\arguments{
\item{user}{Github user name.}

\item{password}{Github password.}
}
\value{
Integer status code of the authentication response from the Github 
API; \code{200} if authentication succesfull.
}
\description{
Authenticate via Github
}

