% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate via Github}
\usage{
authenticate(user, password)
}
\arguments{
\item{user}{character. Github user name.}

\item{password}{character. Github password.}
}
\value{
integer. Status code of the response: \code{200} if
authentication succesfull.
}
\description{
Authenticate via Github
}
